# Speedup NewNonnegativePolynomial for sdsos polynomial