--- a/solvers/mathematical_program.cc
+++ b/solvers/mathematical_program.cc
@@ -786,11 +786,11 @@ MathematicalProgram::AddPositiveDiagonallyDominantMatrixConstraint(
 }
 
 namespace {
-// Add the slack variable for scaled diagonally dominant matrix constraint. The
-// only thing left in AddScaledDiagonallyDominantMatrixConstraint
-// is that the diagonal terms in the sdd matrix should match the summation of
-// the diagonally terms in the slack variable, and the rotated Lorentz cone
-// constraint on M.
+// Add the slack variable for scaled diagonally dominant matrix constraint. In
+// AddScaledDiagonallyDominantMatrixConstraint, we should add the constraint
+// that the diagonal terms in the sdd matrix should match the summation of
+// the diagonally terms in the slack variable, and the upper diagonal corner
+// in M[i][j] should satisfy the rotated Lorentz cone constraint.
 template <typename T>
 void AddSlackVariableForScaledDiagonallyDominantMatrixConstraint(
     const Eigen::Ref<const MatrixX<T>>& X, MathematicalProgram* prog,