--- a/solvers/test/scaled_diagonally_dominant_matrix_test.cc
+++ b/solvers/test/scaled_diagonally_dominant_matrix_test.cc
@@ -10,28 +10,41 @@ using drake::symbolic::test::ExprEqual;
 
 namespace drake {
 namespace solvers {
-GTEST_TEST(ScaledDiagonallyDominantMatrixTest, AddConstraint) {
-  MathematicalProgram prog;
-  auto X = prog.NewSymmetricContinuousVariables<4>();
+bool is_zero(const symbolic::Variable& v) { return v.is_dummy(); }
 
-  auto M = prog.AddScaledDiagonallyDominantMatrixConstraint(
-      X.cast<symbolic::Expression>());
-  for (int i = 0; i < 4; ++i) {
-    for (int j = 0; j < 4; ++j) {
+template <typename T>
+void CheckOffDiagonalTerms(int nx) {
+  MathematicalProgram prog;
+  auto X = prog.NewSymmetricContinuousVariables(nx);
+
+  auto M = prog.AddScaledDiagonallyDominantMatrixConstraint(X.cast<T>());
+  for (int i = 0; i < nx; ++i) {
+    for (int j = 0; j < nx; ++j) {
+      // M[i][j] should be a zero matrix if i >= j.
       if (i >= j) {
         for (int m = 0; m < 2; ++m) {
           for (int n = 0; n < 2; ++n) {
-            EXPECT_EQ(M[i][j](m, n), 0);
+            EXPECT_TRUE(is_zero(M[i][j](m, n)));
           }
         }
       } else {
-        EXPECT_PRED2(ExprEqual, M[i][j](0, 1).Expand(),
+        EXPECT_PRED2(ExprEqual, symbolic::Expression(M[i][j](0, 1)),
                      symbolic::Expression(X(i, j)));
+        EXPECT_PRED2(ExprEqual, symbolic::Expression(M[i][j](1, 0)),
+                     symbolic::Expression(X(j, i)));
       }
     }
   }
 }
 
+GTEST_TEST(ScaledDiagonallyDominantMatrixTest, AddConstraint) {
+  CheckOffDiagonalTerms<symbolic::Expression>(2);
+  CheckOffDiagonalTerms<symbolic::Variable>(2);
+  CheckOffDiagonalTerms<symbolic::Expression>(4);
+  CheckOffDiagonalTerms<symbolic::Variable>(4);
+}
+
+template <typename T>
 void CheckSDDMatrix(const Eigen::Ref<const Eigen::MatrixXd>& X_val,
                     bool is_sdd) {
   // Test if a sdd matrix satisfies the constraint.
@@ -39,8 +52,7 @@ void CheckSDDMatrix(const Eigen::Ref<const Eigen::MatrixXd>& X_val,
   DRAKE_DEMAND(X_val.cols() == nx);
   MathematicalProgram prog;
   auto X = prog.NewSymmetricContinuousVariables(nx);
-  auto M = prog.AddScaledDiagonallyDominantMatrixConstraint(
-      X);
+  auto M = prog.AddScaledDiagonallyDominantMatrixConstraint(X.cast<T>());
 
   for (int i = 0; i < nx; ++i) {
     prog.AddBoundingBoxConstraint(X_val.col(i), X_val.col(i), X.col(i));
@@ -61,13 +73,13 @@ void CheckSDDMatrix(const Eigen::Ref<const Eigen::MatrixXd>& X_val,
     M_sum.setZero();
     for (int i = 0; i < nx; ++i) {
       M_val[i].resize(nx);
-      for (int j = 0; j < nx; ++j) {
+      for (int j = i + 1; j < nx; ++j) {
         M_val[i][j].resize(nx, nx);
         M_val[i][j].setZero();
-        M_val[i][j](i, i) = M[i][j](0, 0).Evaluate(env);
-        M_val[i][j](i, j) = M[i][j](0, 1).Evaluate(env);
+        M_val[i][j](i, i) = symbolic::Expression(M[i][j](0, 0)).Evaluate(env);
+        M_val[i][j](i, j) = symbolic::Expression(M[i][j](0, 1)).Evaluate(env);
         M_val[i][j](j, i) = M_val[i][j](i, j);
-        M_val[i][j](j, j) = M[i][j](1, 1).Evaluate(env);
+        M_val[i][j](j, j) = symbolic::Expression(M[i][j](1, 1)).Evaluate(env);
         M_sum += M_val[i][j];
       }
     }
@@ -112,15 +124,18 @@ GTEST_TEST(ScaledDiagonallyDominantMatrixTest, TestSDDMatrix) {
            0.3, 0.5, 3, 2,
            -0.45, 1, 2, 4;
   // clang-format on
-  CheckSDDMatrix(dd_X, true);
+  CheckSDDMatrix<symbolic::Expression>(dd_X, true);
+  CheckSDDMatrix<symbolic::Variable>(dd_X, true);
 
   Eigen::Matrix4d D = Eigen::Vector4d(1, 2, 3, 4).asDiagonal();
   Eigen::Matrix4d sdd_X = D * dd_X * D;
-  CheckSDDMatrix(sdd_X, true);
+  CheckSDDMatrix<symbolic::Expression>(sdd_X, true);
+  CheckSDDMatrix<symbolic::Variable>(sdd_X, true);
 
   D = Eigen::Vector4d(0.2, -1, -0.5, 1.2).asDiagonal();
   sdd_X = D * dd_X * D;
-  CheckSDDMatrix(sdd_X, true);
+  CheckSDDMatrix<symbolic::Expression>(sdd_X, true);
+  CheckSDDMatrix<symbolic::Variable>(sdd_X, true);
 
   // not_dd_X is not diagonally dominant (dd), but is scaled diagonally
   // dominant.
@@ -128,7 +143,8 @@ GTEST_TEST(ScaledDiagonallyDominantMatrixTest, TestSDDMatrix) {
   not_dd_X << 1, -0.2, 0.3, -0.55, -0.2, 2, 0.5, 1, 0.3, 0.5, 3, 2, -0.55, 1, 2,
       4;
   DRAKE_DEMAND(IsMatrixSDD(not_dd_X));
-  CheckSDDMatrix(not_dd_X, true);
+  CheckSDDMatrix<symbolic::Expression>(not_dd_X, true);
+  CheckSDDMatrix<symbolic::Variable>(not_dd_X, true);
 }
 
 GTEST_TEST(ScaledDiagonallyDominantMatrixTest, TestNotSDDMatrix) {
@@ -141,15 +157,18 @@ GTEST_TEST(ScaledDiagonallyDominantMatrixTest, TestNotSDDMatrix) {
                -1.55, 1, 2, 4;
   // clang-format on
   DRAKE_DEMAND(!IsMatrixSDD(not_sdd_X));
-  CheckSDDMatrix(not_sdd_X, false);
+  CheckSDDMatrix<symbolic::Expression>(not_sdd_X, false);
+  CheckSDDMatrix<symbolic::Variable>(not_sdd_X, false);
 
   Eigen::Matrix4d D = Eigen::Vector4d(1, 2, 3, 4).asDiagonal();
   not_sdd_X = D * not_sdd_X * D;
-  CheckSDDMatrix(not_sdd_X, false);
+  CheckSDDMatrix<symbolic::Expression>(not_sdd_X, false);
+  CheckSDDMatrix<symbolic::Variable>(not_sdd_X, false);
 
   D = Eigen::Vector4d(0.2, -1, -0.5, 1.2).asDiagonal();
   not_sdd_X = D * not_sdd_X * D;
-  CheckSDDMatrix(not_sdd_X, false);
+  CheckSDDMatrix<symbolic::Expression>(not_sdd_X, false);
+  CheckSDDMatrix<symbolic::Variable>(not_sdd_X, false);
 }
 
 GTEST_TEST(SdsosTest, SdsosPolynomial) {