When we create an sdsos polynomial calling `MathematicalProgram::NewNonnegativePolynomial` for a large size monomial basis, the speed was very slow. In one of my test for a monomial basis of size 256, it took 200s in `NewNonnegativePolynomial`.

In this PR we speed up `NewNonnegativePolynomial` for sdsos polynomial. Now it takes 0.4s

The tricks are

1. Instead of declaring each new continuous variables in a double for loop as https://github.com/RobotLocomotion/drake/blob/fbc623dafa8e257a93f169d4ffe7962977f92ae0/solvers/mathematical_program.cc#L795-L801
we now declare all continuous variables in a batch as https://github.com/hongkai-dai/drake/blob/be65bfbda743a483420eef15e613a5cbd28cd590/solvers/mathematical_program.cc#L808-L809. This improves the computation time from 200s to 2s.
2. I created an overloaded the function `AddScaledDiagonallyDominantMatrixConstraint(X)` with `X` being a matrix of *symbolic::Variable*, in addition to a matrix of *symbolic::Expression*. This `symbolic::Variable` version is faster than the `Expression` version. This further improves the computation time from 2s to 0.4s.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/robotlocomotion/drake/10242)
<!-- Reviewable:end -->